[kde] Use rp_qsizetype (formerly rp_qstringsizetype) in some more places.
For more things that were changed to use qsizetype in Qt6.

Fixes the following clang-tidy warnings:

src/kde/OptionsMenuButton.cpp:90:29: warning: narrowing conversion from 'qsizetype' (aka 'long long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   90 |                 romOps_firstActionIndex = menuOptions->children().count();
      |                                           ^
src/kde/config/AchievementsItemDelegate.cpp:70:21: warning: narrowing conversion from 'qsizetype' (aka 'long long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   70 |         const int nl_pos = s_ach.indexOf(QChar(L'\n'));
      |                            ^
src/kde/config/AchievementsItemDelegate.cpp:223:21: warning: narrowing conversion from 'qsizetype' (aka 'long long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  223 |         const int nl_pos = s_ach.indexOf(QChar(L'\n'));
      |                            ^
src/kde/rp_create_thumbnail.cpp:218:27: warning: narrowing conversion from 'qsizetype' (aka 'long long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  218 |         const int bytesPerLine = outParams.retImg.bytesPerLine();
      |                                  ^
src/kde/AchQtDBus.cpp:171:19: warning: narrowing conversion from 'qsizetype' (aka 'long long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  171 |                 nis.rowstride = icon.bytesPerLine();
      |                                 ^
src/kde/ListDataSortProxyModel.cpp:25:23: warning: narrowing conversion from 'qsizetype' (aka 'long long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   25 |         const int str_size = str.size();
      |                              ^

+++ src/kde/AchQtDBus.cpp
@@ -168,7 +168,7 @@ int AchQtDBus::notifyFunc(Achievements::ID id)
 		NotifyIconStruct nis;
 		nis.width = icon.width();
 		nis.height = icon.height();
-		nis.rowstride = icon.bytesPerLine();
+		nis.rowstride = static_cast<int>(icon.bytesPerLine());
 		nis.has_alpha = true;
 		nis.bits_per_sample = 8;	// 8 bits per *channel*.
 		nis.channels = 4;
